backend/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
passlib==1.7.4
python-jose==3.4.0
psycopg2-binary==2.9.10
pymongo==3.12.1
stripe==2.66.0
requests==2.28.1
gunicorn==20.1.0
python-multipart==0.0.5
